Added Exceptions for CountryLanguageController and Service Layer. Started Tests for CountryLanguageController

+++ src/main/java/org/example/hibernatehaunters/controller/CountryLanguageController.java
@@ -1,7 +1,10 @@
 package org.example.hibernatehaunters.controller;
 
 
+import org.apache.coyote.BadRequestException;
 import org.example.hibernatehaunters.models.entities.CountryLanguageEntity;
+import org.example.hibernatehaunters.models.exceptions.countrylanguage.CountryLanguageBadRequestException;
+import org.example.hibernatehaunters.models.exceptions.countrylanguage.CountryLanguageNotFoundException;
 import org.example.hibernatehaunters.service.CountryLanguageService;
 import org.example.hibernatehaunters.service.WorldService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,6 +16,8 @@ import java.util.List;
 public class CountryLanguageController {
     private final CountryLanguageService countryLanguageService;
     private final WorldService worldService;
+
+
 
     @Autowired
     public CountryLanguageController(CountryLanguageService countryLanguageService, WorldService worldService) {
@@ -26,30 +31,43 @@ public class CountryLanguageController {
 
 
     @GetMapping("/countryLanguage")
-    public List<CountryLanguageEntity> getCountryLanguageByCountry(@RequestParam String code) {
+    public List<CountryLanguageEntity> getCountryLanguageByCountry(@RequestParam String code) throws CountryLanguageNotFoundException {
+        List<CountryLanguageEntity> countryLanguageEntities = countryLanguageService.getCountryLanguageByCountryCode(code);
+        if (countryLanguageEntities.isEmpty()){
+            throw new CountryLanguageNotFoundException("With code: " + code);
+        }
         return countryLanguageService.getCountryLanguageByCountryCode(code);
     }
 
+
     //create
     @PostMapping("/countryLanguage/create")
-    public CountryLanguageEntity addLanguage(@RequestBody CountryLanguageEntity countryLanguage){
-        countryLanguageService.createCountryLanguage(countryLanguage);
+    public CountryLanguageEntity addLanguage(@RequestBody CountryLanguageEntity countryLanguage) throws CountryLanguageBadRequestException {
+
+        try {
+            countryLanguageService.createCountryLanguage(countryLanguage);
+        } catch(Exception e){
+            throw new CountryLanguageBadRequestException("Country language creation failed: " + e.getMessage());
+        }
         return countryLanguage;
     }
-
+    //update
     @PutMapping("/countryLanguage/update/official")
-    public CountryLanguageEntity updateLanguageIsOfficial(@RequestBody CountryLanguageEntity countryLanguage){
+    public CountryLanguageEntity updateLanguageIsOfficial(@RequestBody CountryLanguageEntity countryLanguage) throws CountryLanguageNotFoundException {
         CountryLanguageEntity countryLanguageUpdated = countryLanguageService.getCountryLanguageById(countryLanguage.getId());
         countryLanguageService.updateCountryLanguage(countryLanguageUpdated.getId(), countryLanguage);
+
         return countryLanguageUpdated;
     }
 
 
     //delete
-    @DeleteMapping("/countryLanguages/delete/country")
-    public List<CountryLanguageEntity> deleteCountryLanguagesByCountryCode(@RequestParam String code)
-    {
+    @DeleteMapping("/countryLanguages/delete")
+    public List<CountryLanguageEntity> deleteCountryLanguagesByCountryCode(@RequestParam String code) throws CountryLanguageNotFoundException {
         List<CountryLanguageEntity> countryLanguageEntities = countryLanguageService.getCountryLanguageByCountryCode(code);
+        if(countryLanguageEntities.isEmpty()){
+            throw new CountryLanguageNotFoundException("with code: " + code);
+        }
         for(CountryLanguageEntity ce : countryLanguageEntities)
         {
             countryLanguageService.deleteCountryLanguage(ce.getId());
@@ -57,17 +75,17 @@ public class CountryLanguageController {
         return countryLanguageEntities;
     }
 
-
-    @DeleteMapping("/countryLanguages/delete/language")
-    public List<CountryLanguageEntity> deleteCountryLanguage (@PathVariable String name){
-        List<CountryLanguageEntity> countryLanguageEntities = countryLanguageService.getCountryLanguageByLanguage(name);
-        for(CountryLanguageEntity ce : countryLanguageEntities)
-        {
-            countryLanguageService.deleteCountryLanguage(ce.getId());
-        }
-        return countryLanguageEntities;
-        
-    }
+//
+//    @DeleteMapping("/countryLanguages/delete/language")
+//    public List<CountryLanguageEntity> deleteCountryLanguage (@PathVariable String name){
+//        List<CountryLanguageEntity> countryLanguageEntities = countryLanguageService.getCountryLanguageByLanguage(name);
+//        for(CountryLanguageEntity ce : countryLanguageEntities)
+//        {
+//            countryLanguageService.deleteCountryLanguage(ce.getId());
+//        }
+//        return countryLanguageEntities;
+//
+//    }
 
 
 }
